Moved all utility functions into their own directory.
