+++ src/crc_scripts/utils/snap_utils.py
